Rename S3-flavored storage naming to object-store and rewrite the substrate docs (multi-substrate plan 07)
The 21 hardcoded "s3 <op>" tracing strings in the storage gateway become
"object-store <op>" (they actively lied on a GCS or Azure deployment), the
six s3_key() key constructors and their ~380 call sites become
object_store_key(), and the endpoint-parse errors name object-store URLs.
Metrics keep their already-substrate-neutral zeppelin_storage_* names and
the s3_* config fields keep theirs, since both really are what they say.
Riding along: ZeppelinStore::head() returns the seam's own ListedObject
instead of leaking object_store::ObjectMeta through the public API, and the
deprecated new_with_native_batch_delete constructor is gone in favor of
new_with_capabilities. Docs: src/storage/CLAUDE.md is rewritten around the
capability matrix (its three-variant claim, stale line pointers, and the
release-vs-renew misattribution are gone; the owned-key family list is the
current 19), the root CLAUDE.md marks the multi-substrate track executed
and lists the new TEST_BACKEND values, tests/CLAUDE.md documents both
emulators, the security guide words durable audit by capability, and
README/CHANGELOG state the honest status: GCS and Azure implemented behind
emulator gates, no real-cloud run yet.

+++ tests/adversarial/artifacts.rs
@@ -654,7 +654,7 @@ async fn capture_head_only(
     match store.head(key).await {
         Ok(meta) => S3ObjectMeta {
             key: key.to_string(),
-            size: Some(meta.size as u64),
+            size: Some(meta.size),
             captured: false,
             missing: false,
             error: None,
